requirements.txt
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=5.0.0